Try to kill the live server process gracefully on test teardown

+++ pytest_flask/fixtures.py
@@ -4,6 +4,9 @@ import time
 import multiprocessing
 import pytest
 import socket
+import signal
+import os
+import logging
 
 try:
     from urllib2 import urlopen
@@ -49,9 +52,10 @@ class LiveServer(object):
     :param port: The port to run application.
     """
 
-    def __init__(self, app, port):
+    def __init__(self, app, port, clean_stop=False):
         self.app = app
         self.port = port
+        self.clean_stop = clean_stop
         self._process = None
 
     def start(self):
@@ -82,7 +86,20 @@ class LiveServer(object):
     def stop(self):
         """Stop application process."""
         if self._process:
-            self._process.terminate()
+            if self.clean_stop:
+                # We wait a maximum of 5 seconds for the server to terminate cleanly
+                timeout = 5
+                try:
+                    os.kill(self._process.pid, signal.SIGINT)
+                    self._process.join(timeout)
+                except Exception as ex:
+                    logging.error('Failed to join the live server process: %r', ex)
+                finally:
+                    if self._process.is_alive():
+                        # If it's still alive, kill it
+                        self._process.terminate()
+            else:
+                self._process.terminate()
 
     def __repr__(self):
         return '<LiveServer listening at %s>' % self.url()
@@ -123,7 +140,8 @@ def live_server(request, app, monkeypatch):
     monkeypatch.setitem(app.config, 'SERVER_NAME',
                         _rewrite_server_name(server_name, str(port)))
 
-    server = LiveServer(app, port)
+    clean_stop = request.config.getvalue('live_server_clean_stop')
+    server = LiveServer(app, port, clean_stop)
     if request.config.getvalue('start_live_server'):
         server.start()
